move shutdown hook thread to JetInstance
