+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/JetInstanceImpl.java
@@ -35,7 +35,6 @@ import java.util.concurrent.Future;
 
 import static com.hazelcast.jet.impl.util.ExceptionUtil.peel;
 import static com.hazelcast.jet.impl.util.Util.uncheckCall;
-import static com.hazelcast.spi.properties.GroupProperty.SHUTDOWNHOOK_POLICY;
 import static java.util.stream.Collectors.toList;
 
 /**
@@ -44,14 +43,11 @@ import static java.util.stream.Collectors.toList;
 public class JetInstanceImpl extends AbstractJetInstance {
     private final NodeEngine nodeEngine;
     private final JetConfig config;
-    private final Thread shutdownHookThread;
 
     public JetInstanceImpl(HazelcastInstanceImpl hazelcastInstance, JetConfig config) {
         super(hazelcastInstance);
         this.nodeEngine = hazelcastInstance.node.getNodeEngine();
         this.config = config;
-        this.shutdownHookThread = shutdownHookThread(nodeEngine);
-        Runtime.getRuntime().addShutdownHook(shutdownHookThread);
     }
 
     @Nonnull @Override
@@ -113,17 +109,7 @@ public class JetInstanceImpl extends AbstractJetInstance {
         JetService jetService = nodeEngine.getService(JetService.SERVICE_NAME);
         jetService.shutDownJobs();
         super.shutdown();
-        Runtime.getRuntime().removeShutdownHook(shutdownHookThread);
     }
 
-    private Thread shutdownHookThread(NodeEngine nodeEngine) {
-        return new Thread(() -> {
-            String policy = nodeEngine.getProperties().getString(SHUTDOWNHOOK_POLICY);
-            if (policy.equals("TERMINATE")) {
-                getHazelcastInstance().getLifecycleService().terminate();
-            } else {
-                shutdown();
-            }
-        });
-    }
+
 }
